requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
pytest>=7.4.0
google-generativeai>=0.3.0
openai>=1.0.0
beautifulsoup4>=4.12.0
